T4T pres check is  implemented using ISO-DEP presence check nak command in NCI2.0
when T4T tag is detected ,  T4T NAK command is sent for presence check.
The response is sent which indicates command is sent successfully or not. When the ntf is
received , if card is present in field the status is NFA_STATUS_Ok otherwise the status is
NFA_STATUS_FAILED.It can be configured from libnfc-brcm file.

Test: compiles.

Change-Id: I58abff57a5a2514c78046f4a736246568382099f

Conflicts:
	src/nfc/nfc/nfc_main.c