Update patch set 5
Patch Set 5: Presubmit-Verified+1

=== Finished presubmit run: L67700000182251720 ===

Status: https://android-build.googleplex.com/builds/where-is-my-cl/android-review/556705?revision=5&workplanId=L67700000182251720




TreeHugger finished with: 13 passed.





WARNING: The tip-of-tree(4839257) for git_master/gce_x86_phone-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4838916), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.









WARNING: The tip-of-tree(4839257) for git_master/marlin-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4838916), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.





WARNING: The tip-of-tree(4839239) for git_master/sailfish-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4838916), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



This is some super-exciting work! Well done!

Have a question? Found a bug? http://go/treehugger-troubleshoot

Patch-set: 5
Tag: autogenerated:TreeHugger:Verify-Pass
Reviewer: Gerrit User 1062513 <1062513@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Verified=+1