Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Any issue met with this setting or simply set to default value?
> A smaller value could support fast error recovery determination of a lost target.

Some issues can be observed with few phones that does not send ATN PDU when a timeout occurs while it is initiator. Thus the transaction fails.
A solution could be to modify NFCDEP_OP to ask the Target to send RTOX, but it is more straightforward to use the NCI2.0 default value of LN_WT.
Otherwise we can put LN_WT = 8 if you prefer, as many phones in the market.

Patch-set: 1
Reviewer: Gerrit User 1151328 <1151328@85c56323-6fa9-3386-8a01-6480fb634889>