Add empty aid as the last entry of aid routing
If App1ication registers AIDs with route location as host and if default
AID route(Empty Aid)  is set to offHost, there is possibility that
default AID route (empty AID) is first entry in the routing table.
With this listen mode routing table configuration, According to
NCI2.0(6.3.5 AID-based Route Selection Process), if reader sends actual
AID which is registered for HOST it will be routed to offHost as the
empty AID matches with all the AIDs.

Fix: empty AID is set as last AID entry. New entry in control block(ECB)
for empty AID is added .

Test: Nfc On/Off, CTS, VTS,
      Install App1 with AIDs to Host, Set default AID route to eSE,
      Check CE for HOST AID
Change-Id: I699121c7be14ab9c0cbd51c19a89037b776da0c9