Save last NFCEE command separately
When receiving NCI responses for NFCEE commands, the handle value is
retrieved from nfc_cd.last_cmd.
If another NCI command (not NFCEE) is sent between the last
NFCEE command and its response, this value is lost.
Adding a dedicated field to save the last NFCEE command will prevent it.

Bug: 205290007
Change-Id: I240f1d3fd38b30f83d286404b7ed01d16ebe9e64