Free buffer used to store RF_INTF_ACTIVATED_NTF before allocating
For some tags like T3T or T5T, if the tag is deactivated after
activation and before some data is retrieved (like System Code for T3T
tags), then the buffer used to store the activation data (p_activate_ntf)
is not freed.
Checking if this buffer is not null before allocating it prevents
a buffer leakage.

Change-Id: Ib878d401bf1c3b4e5a13fe49886d1821cd2de15d