Add empty aid as the last entry of aid routing
If Application registers AIDs with route location as host and if default
AID route(Empty Aid)  is set to offHost, there is possibility that
default AID route (empty AID) is first entry in the routing table.
With this listen mode routing table configuration, According to
NCI2.0(6.3.5 AID-based Route Selection Process), if reader sends actual
AID which is registered for HOST it will be routed to offHost as the
empty AID matches with all the AIDs.

Bug: 235804655

Test: Nfc On/Off, CTS, VTS, Install App with AIDs to Host,
Set default AID route to eSE, Check CE for HOST AID

Change-Id: Ibb87ac1deb3ae0d4d2fb3e144ca38afbd78202a4