Set OSTicks to 0 when initializing the NFC task
Found out the times() function returns a number close to 0xFFFFFFFF in the 32-bit system, that make the variable - OSTicks could be easily add-overflow.
After tracing the code, found out there is no need to set OSTicks to the value of the times() function.
Because OSTicks is only used as a counter but not a timestamp indicating
the current time.

Bug: 265802734
Test: Toggle NFC Service to see if add-overflow
Change-Id: I5383919ea21097d33e1a2635a115252eefc3c0ca