Update NCI Version checks to accept future versions
All NCI version checks added for the support of v2.0 in the stack
are written as version == NCI_VERSION_2_0 or
version != NCI_VERSION_2_0.

This change modifies all checks to be version >= NCI_VERSION_2_0
and version < NCI_VERSION_2_0 to ensure that chips supporting
NCI versions v2.1 and v2.2 also pass the checks.

Bug: 294393853
Test: m libnfc-nci
Change-Id: I3d4f4e1027efb995ae3fda2a148cf7585731ab55