Revert "linbfc_nci: Make it available for nfc apex"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f893ea8e-a6c2-4b4d-ac06-1192a75b8424). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9e9829da70cc4a45cbe8dc36396a8f9642a47abd