Revert "linbfc_nci: Address review comments"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/13b75b8a-363e-4ca3-a9e9-5e3c37d04a65). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0cab67d2a3f0b2ae3d2ec871aaaa03c991e59337