Revert "linbfc_nci: Make it available for nfc apex"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/13b75b8a-363e-4ca3-a9e9-5e3c37d04a65). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia87091eb189f2117c14900d8c461a5e806b80e10