Casimir: Implement gRPC server
Implements in a second binary a grpc service that allows
querying and modifying the position of devices managed
by casimir.

Two devices may exchange RF packets only if close in contact.

The default binary is unchanged: the positions are not managed
and RF packets are received by all created devices.

Bug: 296109617
Test: cargo build --bin casimir-grpc
Change-Id: I7c329f9894d5faf80d525779daf5030b741954fe