Revert "Keep legacy T5T tags readable in the field"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/caef7a07-1010-4c25-97e3-57ff38ca5190). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I46263553db88e677a1f473ed35e34dfc99bb7e2b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com