Revert "Keep legacy T5T tags readable in the field"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3f079e32-7d15-40d8-9675-a35cf440a264). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia7752857916ed91756e3c7b94be6c47a6b099626
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com