Revert "Add new aconfig dependencies"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fe1c635a-c16a-4ed3-996f-03034f92aa0d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia7aaef9a9bea1e362924d61e403a951073386336
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com