Revert "Keep legacy T5T tags readable in the field"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4da16583-fbe7-4dbf-b989-2500ec8dd25b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0aeb37bd19fd1bfeab4ce42565b396f6b169d514
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com